Update index.js
Exercise 2.3 Course contents step 8. Now the application will work regardless of whether a new part is added or removed. That is, the Total component now takes care of the exercises total automatically after adding or removing a part. of the course. Also I have replaced [ ...course.parts ] with course.parts in the Content component.